modules/data_warehouse/assets/bigquery-external-function/requirements.txt
google-cloud-bigquery==3.5.0
google-cloud-storage==2.7.0